tidy up: split up directories finished
