+++ demos/demo_rainbow.py
@@ -32,16 +32,16 @@ from distutils.version import StrictVersion
 from pyrateoptics.sampling2d import raster
 from pyrateoptics.material.material_isotropic import ConstantIndexGlass
 from pyrateoptics.material.material_glasscat import refractiveindex_dot_info_glasscatalog, CatalogMaterial
-from pyrateoptics.core.surfShape import Asphere
-from pyrateoptics.core.optical_element import OpticalElement
-from pyrateoptics.core.surface import Surface
-from pyrateoptics.core.optical_system import OpticalSystem
-from pyrateoptics.core.ray import RayBundle
+from pyrateoptics.raytracer.surfShape import Asphere
+from pyrateoptics.raytracer.optical_element import OpticalElement
+from pyrateoptics.raytracer.surface import Surface
+from pyrateoptics.raytracer.optical_system import OpticalSystem
+from pyrateoptics.raytracer.ray import RayBundle
 
-from pyrateoptics.core.aperture import CircularAperture
-from pyrateoptics.core.localcoordinates import LocalCoordinates
+from pyrateoptics.raytracer.aperture import CircularAperture
+from pyrateoptics.raytracer.localcoordinates import LocalCoordinates
 
-from pyrateoptics.core.globalconstants import canonical_ey, degree
+from pyrateoptics.raytracer.globalconstants import canonical_ey, degree
 
 import math
 import logging
