+++ demos/demo_zmx.py
@@ -30,10 +30,10 @@ import sys
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
-from pyrateoptics.core.localcoordinates import LocalCoordinates
+from pyrateoptics.raytracer.localcoordinates import LocalCoordinates
 from pyrateoptics.material.material_isotropic import ConstantIndexGlass
-from pyrateoptics.core.globalconstants import standard_wavelength, numerical_tolerance
-from pyrateoptics.core.ray import RayBundle
+from pyrateoptics.raytracer.globalconstants import standard_wavelength, numerical_tolerance
+from pyrateoptics.raytracer.ray import RayBundle
 from pyrateoptics.io.zmx import ZMXParser
 
 from pyrateoptics.sampling2d import raster
