+++ pyrateoptics/material/material_isotropic.py
@@ -25,12 +25,15 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
 import numpy as np
-from ..core.ray import RayBundle
-from ..core.helpers_math import checkfinite
-from ..optimize.optimize import OptimizableVariable
+
+from ..core.base import OptimizableVariable
+
+from ..raytracer.ray import RayBundle
+from ..raytracer.helpers_math import checkfinite
+from ..raytracer.globalconstants import standard_wavelength
+
 from material import MaxwellMaterial
 
-from ..core.globalconstants import standard_wavelength
 
 class IsotropicMaterial(MaxwellMaterial):
